Move PluginLoader to resotoworker
